+++ packages/app/src/features/easy-borrow/components/form/Deposits.tsx
@@ -3,7 +3,7 @@ import { TokenWithBalance } from '@/domain/common/types'
 import { TokenSymbol } from '@/domain/types/TokenSymbol'
 import { Button, ButtonIcon } from '@/ui/atoms/button/Button'
 import { Link } from '@/ui/atoms/link/Link'
-import { Panel } from '@/ui/atoms/new/panel/Panel'
+import { Panel } from '@/ui/atoms/panel/Panel'
 import { Info } from '@/ui/molecules/info/Info'
 import { AssetInput } from '@/ui/organisms/new/asset-input/AssetInput'
 import { cn } from '@/ui/utils/style'
