+++ packages/app/src/features/easy-borrow/components/form/LoanToValuePanel.tsx
@@ -1,5 +1,5 @@
 import { formatPercentage } from '@/domain/common/format'
-import { Panel } from '@/ui/atoms/new/panel/Panel'
+import { Panel } from '@/ui/atoms/panel/Panel'
 import { testIds } from '@/ui/utils/testIds'
 import { nonZeroOrDefault } from '@/utils/bigNumber'
 import { Percentage } from '@marsfoundation/common-universal'
