+++ packages/app/src/features/farm-details/components/farm-info-panel/active/ActiveFarmInfoPanel.tsx
@@ -7,7 +7,7 @@ import { RewardPointsSyncStatus } from '@/features/farm-details/types'
 import { getTokenImage } from '@/ui/assets'
 import { Button } from '@/ui/atoms/button/Button'
 import { DelayedComponent } from '@/ui/atoms/delayed-component/DelayedComponent'
-import { Panel } from '@/ui/atoms/new/panel/Panel'
+import { Panel } from '@/ui/atoms/panel/Panel'
 import { cn } from '@/ui/utils/style'
 import { testIds } from '@/ui/utils/testIds'
 import { assert } from '@/utils/assert'
