+++ packages/app/src/features/farm-details/components/farm-info-panel/inactive/InactiveFarmInfoPanel.tsx
@@ -5,7 +5,7 @@ import { AssetsGroup, Farm } from '@/domain/farms/types'
 import { USD_MOCK_TOKEN } from '@/domain/types/Token'
 import { Button } from '@/ui/atoms/button/Button'
 import { Link } from '@/ui/atoms/link/Link'
-import { Panel } from '@/ui/atoms/new/panel/Panel'
+import { Panel } from '@/ui/atoms/panel/Panel'
 import { links } from '@/ui/constants/links'
 import { testIds } from '@/ui/utils/testIds'
 import { mainnet } from 'viem/chains'
