+++ packages/app/src/features/market-details/components/my-wallet/MyWalletChainMismatch.tsx
@@ -1,4 +1,4 @@
-import { Panel } from '@/ui/atoms/new/panel/Panel'
+import { Panel } from '@/ui/atoms/panel/Panel'
 
 export function MyWalletChainMismatch() {
   return (
