+++ packages/app/src/features/market-details/components/my-wallet/components/WalletPanelContent.tsx
@@ -1,4 +1,4 @@
-import { Panel } from '@/ui/atoms/new/panel/Panel'
+import { Panel } from '@/ui/atoms/panel/Panel'
 import { ReactNode } from 'react'
 
 interface WalletPanelContentProps {
