+++ packages/app/src/features/market-details/components/oracle-panel/components/YieldingFixedOraclePanel.tsx
@@ -1,6 +1,6 @@
 import { YieldingFixedOracleInfo } from '@/domain/oracles/types'
 import { assets } from '@/ui/assets'
-import { Panel } from '@/ui/atoms/new/panel/Panel'
+import { Panel } from '@/ui/atoms/panel/Panel'
 import { BlockExplorerAddressLink } from '@/ui/molecules/block-explorer-address-link/BlockExplorerAddressLink'
 import { InfoTile } from '@/ui/molecules/info-tile/InfoTile'
 import { Info } from '@/ui/molecules/info/Info'
