+++ packages/app/src/features/market-details/components/status-panel/components/StatusPanelGrid.tsx
@@ -1,4 +1,4 @@
-import { Panel } from '@/ui/atoms/new/panel/Panel'
+import { Panel } from '@/ui/atoms/panel/Panel'
 import { ReactNode } from 'react'
 
 interface StatusPanelGridProps {
