+++ packages/app/src/features/markets/views/MarketsView.tsx
@@ -1,5 +1,5 @@
 import { NetworkBadge } from '@/ui/atoms/network-badge/NetworkBadge'
-import { Panel } from '@/ui/atoms/new/panel/Panel'
+import { Panel } from '@/ui/atoms/panel/Panel'
 import { PageLayout } from '@/ui/layouts/PageLayout'
 import { LabeledSwitch } from '@/ui/molecules/labeled-switch/LabeledSwitch'
 import { testIds } from '@/ui/utils/testIds'
