+++ packages/app/src/features/my-portfolio/components/borrow-table/BorrowTable.tsx
@@ -5,7 +5,7 @@ import { borrowDialogConfig } from '@/features/dialogs/borrow/BorrowDialog'
 import { eModeDialogConfig } from '@/features/dialogs/e-mode/EModeDialog'
 import { repayDialogConfig } from '@/features/dialogs/repay/RepayDialog'
 import { Button } from '@/ui/atoms/button/Button'
-import { Panel } from '@/ui/atoms/new/panel/Panel'
+import { Panel } from '@/ui/atoms/panel/Panel'
 import { ApyTooltip } from '@/ui/molecules/apy-tooltip/ApyTooltip'
 import { ActionsCell } from '@/ui/molecules/data-table/components/ActionsCell'
 import { CompactValueCell } from '@/ui/molecules/data-table/components/CompactValueCell'
