+++ packages/app/src/features/my-portfolio/views/PositionView.tsx
@@ -3,7 +3,7 @@ import { LiquidationDetails } from '@/domain/market-info/getLiquidationDetails'
 import { OpenDialogFunction } from '@/domain/state/dialogs'
 import { PageLayout } from '@/ui/layouts/PageLayout'
 
-import { Panel } from '@/ui/atoms/new/panel/Panel'
+import { Panel } from '@/ui/atoms/panel/Panel'
 import { HealthFactorPanelContent } from '@/ui/molecules/new/health-factor-panel-content/HealthFactorPanelContent'
 import { BorrowTable } from '../components/borrow-table/BorrowTable'
 import { CreatePositionPanel } from '../components/create-position-panel/CreatePositionPanel'
