+++ packages/app/src/features/savings/components/savings-token-panel/SavingsTokenPanel.tsx
@@ -5,7 +5,7 @@ import { OpenDialogFunction } from '@/domain/state/dialogs'
 import { Token } from '@/domain/types/Token'
 import { savingsWithdrawDialogConfig } from '@/features/dialogs/savings/withdraw/SavingsWithdrawDialog'
 import { Button } from '@/ui/atoms/button/Button'
-import { Panel } from '@/ui/atoms/new/panel/Panel'
+import { Panel } from '@/ui/atoms/panel/Panel'
 import { cn } from '@/ui/utils/style'
 import { testIds } from '@/ui/utils/testIds'
 import { Percentage } from '@marsfoundation/common-universal'
