+++ packages/app/src/features/savings/components/upgrade-savings-banner/UpgradeSavingsBanner.tsx
@@ -2,7 +2,7 @@ import { formatPercentage } from '@/domain/common/format'
 import { assets } from '@/ui/assets'
 import { Button } from '@/ui/atoms/button/Button'
 import { LinkButton } from '@/ui/atoms/link-button/LinkButton'
-import { Panel } from '@/ui/atoms/new/panel/Panel'
+import { Panel } from '@/ui/atoms/panel/Panel'
 import { links } from '@/ui/constants/links'
 import { cn } from '@/ui/utils/style'
 import { testIds } from '@/ui/utils/testIds'
