+++ packages/app/src/ui/charts/components/ChartTabsPanel.tsx
@@ -1,6 +1,6 @@
 import { DelayedComponent } from '@/ui/atoms/delayed-component/DelayedComponent'
-import { Panel } from '@/ui/atoms/new/panel/Panel'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/ui/atoms/new/tabs/Tabs'
+import { Panel } from '@/ui/atoms/panel/Panel'
 import { useResizeObserver } from '@/ui/utils/useResizeObserver'
 import { assert } from '@/utils/assert'
 import { AlertTriangle, Loader2 } from 'lucide-react'
